// src/app/interfaces/common/visions-resources.interface.ts
export interface VisionsResources {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  imageEn?: string;
  images?: string[];
  informations?: any[];
  description?: string;
  isHtml?: boolean;
  descriptionEn?: string;
  isHtmlEn?: boolean;
  websiteUrl?: string;
  websiteUrlEn?: string;
  slug?: string;
  pages?: string;
  createdAt?: Date;
  updatedAt?: Date;
}